Convert RLE to use longs for start and end time.  Converting back and forth between int and long is error prone and we have not yet commited to keeping time in seconds for 4.0.

+++ src/org/hyperic/hq/measurement/server/session/AvailabilityDataId.java
@@ -29,18 +29,18 @@ import java.io.Serializable;
 
 public class AvailabilityDataId implements Serializable {
     private static final long serialVersionUID = 1L;
-    private int _startime;
+    private long _startime;
     private Measurement _measurement;
     
     public AvailabilityDataId() {
     }
     
-    public AvailabilityDataId(int startime, Measurement measurement) {
+    public AvailabilityDataId(long startime, Measurement measurement) {
         _startime = startime;
         _measurement = measurement;
     }
 
-    public int getStartime() {
+    public long getStartime() {
         return _startime;
     }
 
@@ -52,8 +52,7 @@ public class AvailabilityDataId implements Serializable {
         _measurement = measurement;
     }
 
-    public void setStartime(int startime) {
+    public void setStartime(long startime) {
         _startime = startime;
     }
-
 }
